Fix detection of keyboard names
Interpretation of input local identifier returned by GetKeyboardLayout was really partial since only the language part was interpreted.
Now the language part is map with names provided by the system through CultureInfo (no need of a Layout.txt mapping file for that).
And the keyboard part is read with more system call and a call to the registry.

So now their is no more limitations on the number of keyboards we can setup in a language, and can switch between multiple keyboards in a same language or in different languages.